requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
aiohttp-socks==0.8.4